backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai==1.3.7
pydantic>=2.10.0,<3
ultralytics>=8.1.0
python-dotenv==1.0.0
pyautogen>=0.7.0
opencv-python>=4.9.0.80
Pillow>=10.1.0
numpy>=1.26.4
imagehash==4.3.1
torch>=2.2.0
dashscope==1.19.1
httpx==0.25.0
pymysql>=1.1.0

